+++ checkpoint/CHANGELOG.md
@@ -19,6 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Support nested `asyncio.run` with `nest_asyncio` library.
+- Support empty tuple values in checkpointable params.
+- Introduce `ListKey` and `TupleKey` for `list` and `tuple` PyTree nodes
+ respectively.
+
 
 ## [0.6.4] - 2024-09-17